CRUD For Cities, Certificates, Departments & Examfees has been done

+++ database/migrations/2021_07_05_11_create_examfees_table.php
@@ -15,7 +15,7 @@ class CreateExamfeesTable extends Migration
     {
         Schema::create('examfees', function (Blueprint $table) {
             $table->id();
-            $table->tinyInteger('price');
+            $table->integer('price');
             $table->foreignId('clinic_id')->constrained()->onUpdate('cascade')->onDelete('cascade');
             $table->timestamps();
         });
